Move legislation under validation
